requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
structlog==23.2.0
python-multipart==0.0.6
pydantic-settings==2.0.3
